[InstCombine] Fix mismatched attribute lists for combined calls
Summary:
Currently, a transformation like pow(2.0, x) -> exp2(x) copies the pow
attribute list verbatim and applies it to exp2. This works out fine
when the attribute list is empty, but when it isn't clang will error due
due to the mismatch.

This was discovered after implementing the `noundef` param attribute.

Reviewers: eugenis, vitalybuka, xbolva00, jdoerfert

Subscribers: lebedev.ri, hiraditya, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D82820