[Attributor] Cost interface for internalizing functions.
This introduces a cost interface to analyse functions that can be
internalized.

Generally, like inlining, internalizing large functions is not
desired, but funtions which can derive very useful IP information
are preferred to be internalized.

So we create a cost interface deciding if the non-exact function
should be internalized or not based on AA dependencies. First we
run fixpoint iterations for some time to get a nearly (if not
already) complete dependency graph, then for each non-exact funtion,
we calculate the cost of internalizing it based on the number of
instructions and the number of AAs that depend on this function.
After that, we clear all the "side effects" produced by the procedure
above and re-generate all AAs and do the actual fixpoint iteration.

One should not that the IP information derived from the cost analysis
period may not be correct after deciding which functions should be
internalized. So all AAs depends on non-exact functions should be
invalid. However, for covinience, we simply delete all AAs and
reidentify them.